ref(ddm): Split utils
