+++ static/app/utils/metrics/dashboard.tsx
@@ -1,8 +1,10 @@
 import {urlEncode} from '@sentry/utils';
 
 import {PageFilters} from 'sentry/types';
-import {emptyWidget, MetricDisplayType, MetricsQuery} from 'sentry/utils/metrics';
+import {emptyWidget} from 'sentry/utils/metrics/constants';
 import {formatMRI, MRIToField} from 'sentry/utils/metrics/mri';
+import type {MetricsQuery} from 'sentry/utils/metrics/types';
+import {MetricDisplayType} from 'sentry/utils/metrics/types';
 import {
   DashboardWidgetSource,
   DisplayType,
