+++ static/app/views/alerts/rules/metric/metricRulePresets.tsx
@@ -2,8 +2,9 @@ import type {LinkProps} from 'sentry/components/links/link';
 import {t} from 'sentry/locale';
 import type {MRI, Project} from 'sentry/types';
 import {DiscoverDatasets, DisplayModes} from 'sentry/utils/discover/types';
-import {getDdmUrl, MetricDisplayType} from 'sentry/utils/metrics';
+import {getDdmUrl} from 'sentry/utils/metrics';
 import {parseField} from 'sentry/utils/metrics/mri';
+import {MetricDisplayType} from 'sentry/utils/metrics/types';
 import type {TimePeriodType} from 'sentry/views/alerts/rules/metric/details/constants';
 import type {MetricRule} from 'sentry/views/alerts/rules/metric/types';
 import {isCustomMetricField} from 'sentry/views/alerts/rules/metric/utils/isCustomMetricField';
