+++ static/app/views/alerts/rules/metric/mriField.tsx
@@ -6,7 +6,8 @@ import Tag from 'sentry/components/tag';
 import {t} from 'sentry/locale';
 import {space} from 'sentry/styles/space';
 import {MetricMeta, MRI, ParsedMRI, Project} from 'sentry/types';
-import {getReadableMetricType, isAllowedOp} from 'sentry/utils/metrics';
+import {isAllowedOp} from 'sentry/utils/metrics';
+import {getReadableMetricType} from 'sentry/utils/metrics/formatters';
 import {
   DEFAULT_METRIC_ALERT_FIELD,
   formatMRI,
