+++ static/app/views/dashboards/utils.tsx
@@ -41,12 +41,9 @@ import {
 } from 'sentry/utils/discover/fields';
 import {DiscoverDatasets, DisplayModes} from 'sentry/utils/discover/types';
 import {getMeasurements} from 'sentry/utils/measurements/measurements';
-import {
-  getDdmUrl,
-  getMetricDisplayType,
-  MetricWidgetQueryParams,
-} from 'sentry/utils/metrics';
+import {getDdmUrl, getMetricDisplayType} from 'sentry/utils/metrics';
 import {parseField} from 'sentry/utils/metrics/mri';
+import type {MetricWidgetQueryParams} from 'sentry/utils/metrics/types';
 import {decodeList} from 'sentry/utils/queryString';
 import theme from 'sentry/utils/theme';
 import {
