+++ static/app/views/ddm/constants.tsx
@@ -1,10 +1,3 @@
-import {SortState} from 'sentry/utils/metrics';
-
 export const DDM_CHART_GROUP = 'ddm_chart_group';
 
 export const MIN_WIDGET_WIDTH = 400;
-
-export const DEFAULT_SORT_STATE: SortState = {
-  name: undefined,
-  order: 'asc',
-};
