+++ static/app/views/ddm/context.tsx
@@ -10,19 +10,18 @@ import * as Sentry from '@sentry/react';
 import isEqual from 'lodash/isEqual';
 
 import {
-  emptyWidget,
   getAbsoluteDateTimeRange,
   getDefaultMetricDisplayType,
-  MetricWidgetQueryParams,
   useInstantRef,
   useUpdateQuery,
 } from 'sentry/utils/metrics';
+import {DEFAULT_SORT_STATE, emptyWidget} from 'sentry/utils/metrics/constants';
+import {MetricWidgetQueryParams} from 'sentry/utils/metrics/types';
 import {useMetricsMeta} from 'sentry/utils/metrics/useMetricsMeta';
 import {decodeList} from 'sentry/utils/queryString';
 import {useLocalStorageState} from 'sentry/utils/useLocalStorageState';
 import usePageFilters from 'sentry/utils/usePageFilters';
 import useRouter from 'sentry/utils/useRouter';
-import {DEFAULT_SORT_STATE} from 'sentry/views/ddm/constants';
 import {FocusArea} from 'sentry/views/ddm/focusArea';
 import {useStructuralSharing} from 'sentry/views/ddm/useStructuralSharing';
