+++ static/app/views/ddm/metricsExplorer.tsx
@@ -1,7 +1,7 @@
 import {useState} from 'react';
 
 import {MRI} from 'sentry/types';
-import {MetricDisplayType, MetricWidgetQueryParams} from 'sentry/utils/metrics';
+import {MetricDisplayType, MetricWidgetQueryParams} from 'sentry/utils/metrics/types';
 import {MetricWidget} from 'sentry/views/ddm/widget';
 
 // TODO(ddm): move this to admin
