+++ static/app/views/ddm/widgetDetails.tsx
@@ -5,7 +5,8 @@ import {TabList, TabPanels, Tabs} from 'sentry/components/tabs';
 import {Tooltip} from 'sentry/components/tooltip';
 import {t} from 'sentry/locale';
 import {space} from 'sentry/styles/space';
-import {isCustomMetric, MetricWidgetQueryParams} from 'sentry/utils/metrics';
+import {isCustomMetric} from 'sentry/utils/metrics';
+import {MetricWidgetQueryParams} from 'sentry/utils/metrics/types';
 import {CodeLocations} from 'sentry/views/ddm/codeLocations';
 import {useDDMContext} from 'sentry/views/ddm/context';
 import {SampleTable} from 'sentry/views/ddm/sampleTable';
