+++ static/app/views/settings/projectMetrics/projectMetrics.tsx
@@ -15,7 +15,8 @@ import {DEFAULT_DEBOUNCE_DURATION} from 'sentry/constants';
 import {t, tct} from 'sentry/locale';
 import {space} from 'sentry/styles/space';
 import {Organization, Project} from 'sentry/types';
-import {getReadableMetricType, METRICS_DOCS_URL} from 'sentry/utils/metrics';
+import {METRICS_DOCS_URL} from 'sentry/utils/metrics/constants';
+import {getReadableMetricType} from 'sentry/utils/metrics/formatters';
 import {formatMRI} from 'sentry/utils/metrics/mri';
 import {useMetricsMeta} from 'sentry/utils/metrics/useMetricsMeta';
 import {middleEllipsis} from 'sentry/utils/middleEllipsis';
