+++ static/app/views/settings/projectMetrics/projectMetricsDetails.tsx
@@ -15,8 +15,10 @@ import SentryDocumentTitle from 'sentry/components/sentryDocumentTitle';
 import {CHART_PALETTE} from 'sentry/constants/chartPalette';
 import {t} from 'sentry/locale';
 import {MetricsOperation, MetricType, MRI, Organization, Project} from 'sentry/types';
-import {getDdmUrl, getReadableMetricType, MetricDisplayType} from 'sentry/utils/metrics';
+import {getDdmUrl} from 'sentry/utils/metrics';
+import {getReadableMetricType} from 'sentry/utils/metrics/formatters';
 import {formatMRI, formatMRIField, MRIToField, parseMRI} from 'sentry/utils/metrics/mri';
+import {MetricDisplayType} from 'sentry/utils/metrics/types';
 import {useMetricsData} from 'sentry/utils/metrics/useMetricsData';
 import {useMetricsTags} from 'sentry/utils/metrics/useMetricsTags';
 import routeTitleGen from 'sentry/utils/routeTitle';
